refactor(Vsix): match namespace to "Json/Completion" folder.
