+++ src/LibraryManager.Vsix/Json/Completion/LibraryIdCompletionProvider.cs
@@ -12,11 +12,10 @@ using Microsoft.VisualStudio.Text;
 using Microsoft.VisualStudio.Utilities;
 using Microsoft.Web.LibraryManager.Contracts;
 using Microsoft.Web.LibraryManager.Vsix.Contracts;
-using Microsoft.Web.LibraryManager.Vsix.Json;
 using Microsoft.WebTools.Languages.Json.Editor.Completion;
 using Microsoft.WebTools.Languages.Json.Parser.Nodes;
 
-namespace Microsoft.Web.LibraryManager.Vsix
+namespace Microsoft.Web.LibraryManager.Vsix.Json.Completion
 {
     [Export(typeof(IJsonCompletionListProvider))]
     [Name(nameof(LibraryIdCompletionProvider))]
