+++ src/LibraryManager.Vsix/Json/Completion/SimpleCompletionEntry.cs
@@ -8,7 +8,7 @@ using Microsoft.VisualStudio.Text;
 using Microsoft.WebTools.Languages.Json.Editor.Completion;
 using Microsoft.WebTools.Languages.Shared.Editor.Completion;
 
-namespace Microsoft.Web.LibraryManager.Vsix
+namespace Microsoft.Web.LibraryManager.Vsix.Json.Completion
 {
     internal class SimpleCompletionEntry : JsonCompletionEntry
     {
